Widen the inter-engine variable slots, and enumerate the rest
Re-enabling the pointer/int conversion warnings the Makefile suppresses turns
up 57 sites. They split in two: pointers that round-trip through a 32-bit
integer, which genuinely break, and addresses truncated only for comparison or
arithmetic, which do not.

Fixed on the round-trip side:

* interEngineVariableSpace was s32. Engines park pointers there and read them
  back — remix7_start_last_remix1_bgm leaves a SoundPlayer * for two other
  functions to pick up — so the top half was being dropped.

* src/code_08047e50.c never included gameplay.h, leaving those same helpers
  implicitly declared as returning int. Widening the header alone would not
  have helped there.

57 -> 54 sites, round-trips 7 -> 5.

The remaining five are a different problem, not a missing cast: bitmap_font.c
and graphics_table.c use bit 31 of a pointer as an indirection flag and mask it
off with & 0x7fffffff. That works when addresses fit in 32 bits. Fixing it
means deciding where the flag should live on a 64-bit host, so it is left
identified rather than half-changed.

+++ src/code_08047e50.c
@@ -1,5 +1,9 @@
 #include "global.h"
 #include "code_08047e50.h"
+// Without this the gameplay_* helpers are implicitly declared as returning
+// int, which truncates the pointers this file parks in and retrieves from the
+// inter-engine variables.
+#include "src/scenes/gameplay.h"
 
 
 /* MISCELLANEOUS SCRIPT FUNCTIONS */
@@ -56,7 +60,7 @@ void remix7_start_last_remix1_bgm(void) {
 
     soundPlayer = play_sound_in_player(MUSIC_PLAYER_1, &Lesson1_seqData);
     set_soundplayer_speed(soundPlayer, INT_TO_FIXED(1.08984375));
-    gameplay_set_inter_engine_variable(0, (s32)soundPlayer);
+    gameplay_set_inter_engine_variable(0, (intptr_t)soundPlayer);
 }
